Make LayerTensors a type alias.
Also move LayerTensors and TensorWrap to a separate module 'tensor'.

The motivation to keep the name LayerTensors is that writing
EnumMap<Layer, TensorWrap<T>> in various places is a little verbose.
