+++ dpar/src/models/tensorflow/mod.rs
@@ -5,3 +5,7 @@ pub use self::guide::*;
 
 mod model;
 pub use self::model::*;
+
+mod tensor;
+pub use self::tensor::LayerTensors;
+pub(crate) use self::tensor::*;
